+++ src/components/AppProviders.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import {AuthProvider} from "../providers/auth";
 import {NavigationProvider} from "../providers/navigation";
 import {FormProvider} from "../providers/form";
 import {NoteListProvider} from "../providers/note";
@@ -7,13 +8,15 @@ import {NoteServiceProvider} from "../providers/noteService";
 function AppProviders({ children }) {
   return (
     <NavigationProvider>
-      <FormProvider>
-        <NoteListProvider>
-          <NoteServiceProvider>
-            {children}
-          </NoteServiceProvider>
-        </NoteListProvider>
-      </FormProvider>
+      <AuthProvider>
+        <FormProvider>
+          <NoteListProvider>
+            <NoteServiceProvider>
+              {children}
+            </NoteServiceProvider>
+          </NoteListProvider>
+        </FormProvider>
+      </AuthProvider>
     </NavigationProvider>
   );
 }
