+++ src/providers/auth.js
@@ -1,6 +1,6 @@
 import React, {useReducer} from "react";
-import {authInitialState, authReducer} from "../../store/reducers/auth";
-import {AuthContext} from "../../store/modules/AuthState";
+import {authInitialState, authReducer} from "../store/reducers/auth";
+import {AuthContext} from "../store/modules/AuthState";
 
 export const AuthProvider = ({ children }) => {
     const [state, dispatch] = useReducer(authReducer, authInitialState)
